Upgrade to v13

+++ submodules/mutes.ts
@@ -1,19 +1,22 @@
 import { PrismaClient } from '@prisma/client';
 const prisma = new PrismaClient();
-let util_functions = require('../util_functions');
+import Discord from 'discord.js';
+import * as util_functions from '../util_functions';
 import * as Types from '../types';
-let setupmute = {
+const setupmute = {
   name: 'setupmute',
   syntax: 'setupmute',
   explanation: 'Configure mutes',
   long_explanation:
     'Create/Delete the mute role, and fix its channel permissions if they get changed',
-  permissions: (msg) => msg.member.hasPermission('MANAGE_ROLES'),
-  responder: async (msg) => {
+  permissions: (msg: Discord.Message) =>
+    msg.member?.permissions.has('MANAGE_ROLES'),
+  responder: async (msg: util_functions.EMessage) => {
+    if (!msg.guild) return;
     if (
       !(await prisma.mute_roles.findFirst({ where: { server: msg.guild.id } }))
     ) {
-      let res = await util_functions.embed_options(
+      const res = await util_functions.embed_options(
         'What do you want to setup?',
         ['Create Mute Role'],
         ['🔨'],
@@ -25,7 +28,7 @@ let setupmute = {
           'MANAGE_CHANNELS',
         ]);
         await msg.channel.send('What should the mute role be named?');
-        let rolename = await msg.channel.awaitMessages(
+        const rolename = await msg.channel.awaitMessages(
           (m) => m.author.id == msg.author.id,
           {
             max: 1,
@@ -39,9 +42,7 @@ let setupmute = {
         let mute_role;
         try {
           mute_role = await msg.guild.roles.create({
-            data: {
-              name: rolename.array()[0].content,
-            },
+            name: rolename.array()[0].content,
             reason: 'Created mute role',
           });
         } catch (e) {
@@ -53,7 +54,7 @@ let setupmute = {
         await msg.channel.send(
           util_functions.desc_embed('Setting channel overrides')
         );
-        let guild_channels = msg.guild.channels.cache.array();
+        const guild_channels = msg.guild.channels.cache.array();
         for (const channel of guild_channels) {
           try {
             await channel.updateOverwrite(mute_role, {
@@ -80,7 +81,7 @@ let setupmute = {
         await Types.LogChannel.tryToLog(msg, `Created mute role ${mute_role}`);
       }
     } else {
-      let res = await util_functions.embed_options(
+      const res = await util_functions.embed_options(
         'What do you want to setup?',
         ['Remove Mute Role', 'Fix Mute Role Channel Overrides'],
         ['🗑️', '🔨'],
@@ -88,16 +89,16 @@ let setupmute = {
       );
       if (res === 0) {
         util_functions.assertHasPerms(msg.guild, ['MANAGE_ROLES']);
-        let conf = await util_functions.confirm(msg);
+        const conf = await util_functions.confirm(msg);
         if (conf) {
-          let mute_role_db = await prisma.mute_roles.findFirst({
+          const mute_role_db = await prisma.mute_roles.findFirst({
             where: {
               server: msg.guild.id,
             },
           });
-          let mute_role = msg.guild.roles.cache.get(mute_role_db.role);
+          const mute_role = msg.guild.roles.cache.get(mute_role_db!.role);
           try {
-            await mute_role.delete();
+            if (mute_role) await mute_role.delete();
           } catch (e) {}
           await prisma.mute_roles.deleteMany({
             where: {
@@ -114,16 +115,16 @@ let setupmute = {
         }
       } else if (res === 1) {
         util_functions.assertHasPerms(msg.guild, ['MANAGE_CHANNELS']);
-        let mute_role_db = await prisma.mute_roles.findFirst({
+        const mute_role_db = await prisma.mute_roles.findFirst({
           where: {
             server: msg.guild.id,
           },
         });
-        let mute_role = msg.guild.roles.cache.get(mute_role_db.role);
-        let guild_channels = msg.guild.channels.cache.array();
+        const mute_role = msg.guild.roles.cache.get(mute_role_db!.role);
+        const guild_channels = msg.guild.channels.cache.array();
         for (const channel of guild_channels) {
           try {
-            await channel.updateOverwrite(mute_role, { SEND_MESSAGES: false });
+            await channel.updateOverwrite(mute_role!, { SEND_MESSAGES: false });
           } catch (e) {
             await msg.channel.send(
               util_functions.desc_embed(
@@ -143,7 +144,10 @@ let setupmute = {
     }
   },
 };
-function checkChannelsThingCanTalkIn(guild, member) {
+function checkChannelsThingCanTalkIn(
+  guild: Discord.Guild,
+  member: Discord.GuildMember | Discord.Role
+) {
   return guild.channels.cache
     .array()
     .filter(
@@ -152,23 +156,32 @@ function checkChannelsThingCanTalkIn(guild, member) {
         channel.permissionsFor(member).has('VIEW_CHANNEL')
     );
 }
-function checkChannelsThingCanTalkInAlways(guild, thing) {
+function checkChannelsThingCanTalkInAlways(
+  guild: Discord.Guild,
+  thing: Discord.GuildMember | Discord.Role
+) {
   return guild.channels.cache.array().filter((channel) => {
-    let po = channel.permissionOverwrites.get(thing.id);
+    const po = channel.permissionOverwrites.get(thing.id);
     return po ? po.allow.has('SEND_MESSAGES') : false;
   });
 }
-var parse_duration = require('parse-duration');
-let mute = {
+import parse_duration from 'parse-duration';
+const mute = {
   name: 'mute',
   syntax: 'mute <user: user_id> [duration: word]',
   explanation: 'Mute a user',
   long_explanation:
     'Mute a user. [DURATION] is an optional duration in the form `5m`',
-  permissions: (msg) => msg.member.hasPermission('MANAGE_ROLES'),
-  responder: async (msg, cmd, client) => {
+  permissions: (msg: Discord.Message) =>
+    msg.member?.permissions.has('MANAGE_ROLES'),
+  responder: async (
+    msg: util_functions.EMessage,
+    cmd: { user: string; duration: string | undefined },
+    client: Discord.Client
+  ) => {
+    if (!msg.guild) return;
     util_functions.assertHasPerms(msg.guild, ['MANAGE_ROLES']);
-    if (cmd.user === client.user.id) {
+    if (cmd.user === client.user?.id) {
       await msg.dbReply('fuck you');
       await msg.dbReply('<a:dance:759943179175854100>');
       return;
@@ -183,29 +196,29 @@ let mute = {
         },
       })
     ) {
-      let mute_role_db = await prisma.mute_roles.findFirst({
+      const mute_role_db = await prisma.mute_roles.findFirst({
         where: {
           server: msg.guild.id,
         },
       });
-      let mute_role = msg.guild.roles.cache.get(mute_role_db.role);
-      let mutee = msg.guild.members.cache.get(cmd.user);
-      if (mutee.roles.highest.position >= msg.member.roles.highest.position) {
+      const mute_role = msg.guild.roles.cache.get(mute_role_db!.role);
+      const mutee = msg.guild.members.cache.get(cmd.user);
+      if (mutee!.roles.highest.position >= msg.member!.roles.highest.position) {
         await msg.channel.send(
           util_functions.desc_embed(
             'The user you are trying to mute is either above you or at the same level as you. You must be above them to mute'
           )
         );
       } else {
-        mutee.roles.add(mute_role);
+        mutee!.roles.add(mute_role!);
         if (cmd.duration && parse_duration(cmd.duration, 's')) {
           await util_functions.schedule_event(
             {
               type: 'unmute',
               channel: msg.channel.id,
-              user: mutee.id,
+              user: mutee!.id,
               server: msg.guild.id,
-              role: mute_role.id,
+              role: mute_role!.id,
             },
             cmd.duration
           );
@@ -222,7 +235,7 @@ let mute = {
           );
           await Types.LogChannel.tryToLog(msg, `Muted ${mutee} indefinitely`);
         }
-        let userCanTalkIn = checkChannelsThingCanTalkIn(msg.guild, mutee);
+        const userCanTalkIn = checkChannelsThingCanTalkIn(msg.guild, mutee!);
         if (userCanTalkIn.length > 0) {
           if (
             await prisma.alert_channels.findFirst({
@@ -233,52 +246,56 @@ let mute = {
           ) {
             try {
               let reason = '';
-              if (checkChannelsThingCanTalkIn(msg.guild, mute_role).length) {
-                let chans = checkChannelsThingCanTalkIn(msg.guild, mute_role);
+              if (checkChannelsThingCanTalkIn(msg.guild, mute_role!).length) {
+                const chans = checkChannelsThingCanTalkIn(
+                  msg.guild,
+                  mute_role!
+                );
                 reason += `In ${chans.join(
                   ' and '
                 )}, the mute role isn't setup. To fix this, run the command \`m: setupmute\` and select the "Fix Mute Role Channel Overrides" option.\n`;
               }
-              let role_allows = [];
-              for (let role of mutee.roles.cache.array()) {
+              const role_allows = [];
+              for (const role of mutee!.roles.cache.array()) {
                 if (
                   checkChannelsThingCanTalkInAlways(msg.guild, role).length &&
                   role.id !== msg.guild.id
                 ) {
-                  let chans = checkChannelsThingCanTalkInAlways(
+                  const chans = checkChannelsThingCanTalkInAlways(
                     msg.guild,
                     role
                   );
                   role_allows.push({ role: role, chans: chans });
                 }
               }
-              for (let allow of role_allows) {
+              for (const allow of role_allows) {
                 reason += `In ${allow.chans.join(' and ')}, ${
                   allow.role
                 } has a permission override. You can fix this in the channel settings\n`;
               }
-              if (checkChannelsThingCanTalkInAlways(msg.guild, mutee).length) {
-                let chans = checkChannelsThingCanTalkInAlways(msg.guild, mutee);
+              if (checkChannelsThingCanTalkInAlways(msg.guild, mutee!).length) {
+                const chans = checkChannelsThingCanTalkInAlways(
+                  msg.guild,
+                  mutee!
+                );
                 reason += `In ${chans.join(
                   ' and '
                 )}, ${mutee} has a permission override. You can fix this in the channel settings\n`;
               }
-              await msg.guild.channels.cache
-                .get(
-                  await prisma.alert_channels.findFirst({
-                    where: {
-                      server: msg.guild.id,
-                    },
-                  }).channel
-                )
-                .send({
-                  content: `${msg.author}`,
-                  embed: util_functions.desc_embed(
-                    `Warning: Muted user ${mutee} can still talk in ${userCanTalkIn.join(
-                      ' and '
-                    )}.\nReasons:\n${reason}`
-                  ),
-                });
+              await (msg.guild.channels.cache.get(
+                (await prisma.alert_channels.findFirst({
+                  where: {
+                    server: msg.guild.id,
+                  },
+                }))!.channel
+              )! as Discord.TextChannel).send({
+                content: `${msg.author}`,
+                embed: util_functions.desc_embed(
+                  `Warning: Muted user ${mutee} can still talk in ${userCanTalkIn.join(
+                    ' and '
+                  )}.\nReasons:\n${reason}`
+                ),
+              });
             } catch (e) {
               console.log(e);
             }
@@ -294,13 +311,20 @@ let mute = {
     }
   },
 };
-let unmute = {
+const unmute = {
   name: 'unmute',
   syntax: 'unmute <user: user_id>',
   explanation: 'Unmute a user',
   long_explanation: 'Unmute a user',
-  permissions: (msg) => msg.member.hasPermission('MANAGE_ROLES'),
-  responder: async (msg, cmd) => {
+  permissions: (msg: Discord.Message) =>
+    msg.member?.permissions.has('MANAGE_ROLES'),
+  responder: async (
+    msg: util_functions.EMessage,
+    cmd: {
+      user: string;
+    }
+  ) => {
+    if (!msg.guild || !msg.member) return;
     util_functions.assertHasPerms(msg.guild, ['MANAGE_ROLES']);
     if (
       await prisma.mute_roles.findFirst({
@@ -309,13 +333,17 @@ let unmute = {
         },
       })
     ) {
-      let mute_role_db = await prisma.mute_roles.findFirst({
+      const mute_role_db = await prisma.mute_roles.findFirst({
         where: {
           server: msg.guild.id,
         },
       });
-      let mute_role = msg.guild.roles.cache.get(mute_role_db.role);
-      let mutee = msg.guild.members.cache.get(cmd.user);
+      if (!mute_role_db)
+        throw new util_functions.BotError('user', "Couldn't find mute role");
+      const mute_role = msg.guild.roles.cache.get(mute_role_db.role);
+      const mutee = msg.guild.members.cache.get(cmd.user);
+      if (!mutee)
+        throw new util_functions.BotError('user', "Couldn't find mutee");
       if (mutee.roles.highest.position >= msg.member.roles.highest.position) {
         await msg.channel.send(
           util_functions.desc_embed(
@@ -323,6 +351,8 @@ let unmute = {
           )
         );
       } else {
+        if (!mute_role)
+          throw new util_functions.BotError('user', "Couldn't find mute role");
         await mutee.roles.remove(mute_role);
         await msg.channel.send(util_functions.desc_embed(`Unmuted ${mutee}`));
         await Types.LogChannel.tryToLog(msg, `Unmuted ${mutee}`);
@@ -336,14 +366,14 @@ let unmute = {
     }
   },
 };
-exports.onChannelCreate = async (channel) => {
-  let mr = await prisma.mute_roles.findFirst({
+exports.onChannelCreate = async (channel: Discord.TextChannel) => {
+  const mr = await prisma.mute_roles.findFirst({
     where: { server: channel.guild.id },
   });
 
   if (mr) {
-    let mute_role = channel.guild.roles.cache.get(mr.role);
-    channel.updateOverwrite(mute_role, {
+    const mute_role = channel.guild.roles.cache.get(mr.role);
+    channel.updateOverwrite(mute_role!, {
       SEND_MESSAGES: false,
       ADD_REACTIONS: false,
     });
